requirements.txt
alpaca-trade-api
pandas
numpy
scikit-learn
matplotlib
python-dotenv